Fix: QA 진행 결과에 따른 오류 업데이트 (#174)
* Fix: 학생 내보내기 API 오류 수정 (쿼리에 lectureId 정보 조건 추가) #173

* Fix: 수업 리포트 Put API 시, 참여자와 숙제 정보의 변경 여부에 따라 조건 분기 되도록 수정 진행 #173

* Fix: 이미 초대한 학생, 다시 초대할 경우 에러 발생하도록 조건 추가 #173

* Fix: 에러 메세지 출력 정보 추가 #173

* Style: 보내는 문자 오탈자 수정 #173

* Fix: 수업 리포트 전체 수정 API 참여자 & 숙제 정보 변경 여부에 따른 조건 분기 추가 #173

* Style: 임시 출력 코드 삭제 #173

+++ src/main/java/com/selfrunner/gwalit/domain/lesson/service/LessonService.java
@@ -34,6 +34,7 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -92,17 +93,67 @@ public class LessonService {
         Lesson lesson = lessonRepository.findById(lessonId).orElseThrow(() -> new LessonException(ErrorCode.NOT_EXIST_LESSON));
         memberAndLectureRepository.findMemberAndLectureByMemberAndLectureLectureId(member, lesson.getLecture().getLectureId()).orElseThrow(() -> new MemberException(ErrorCode.UNAUTHORIZED_EXCEPTION));
 
-        // Business Logic
-        lesson.update(putLessonReq);
-        homeworkRepository.deleteHomeworkByLessonId(lessonId);
-        List<Homework> homeworkList = new ArrayList<>();
-        for (Participant participant : putLessonReq.getParticipants()) {
-            List<Homework> tempHomeworkList = putLessonReq.getHomeworks().stream()
-                    .map(homeworkReq -> HomeworkReq.staticToEntity(homeworkReq, participant.getMemberId(), lesson.getLessonId()))
-                    .collect(Collectors.toList());
-            homeworkList.addAll(tempHomeworkList);
+        // Business Logic: Homework 변경 여부 확인 진행
+        // 참여자 비교
+        Boolean needUpdate = Boolean.FALSE;
+        if(lesson.getParticipants().size() != putLessonReq.getParticipants().size()) {
+            needUpdate = Boolean.TRUE;
         }
-        homeworkRepository.saveAll(homeworkList);
+        if(lesson.getParticipants().size() == putLessonReq.getParticipants().size()) {
+            for(Participant lessonParticipant : lesson.getParticipants()) {
+                Boolean temp = Boolean.FALSE;
+                for(Participant dtoParticipant : putLessonReq.getParticipants()) {
+                    if(lessonParticipant.getMemberId().equals(dtoParticipant.getMemberId())) {
+                        temp = Boolean.TRUE;
+                    }
+                }
+                if(!temp) {
+                    needUpdate = Boolean.TRUE;
+                    break;
+                }
+            }
+        }
+        // 숙제 비교
+        List<Homework> homeworkRowList = homeworkRepository.findAllByMemberIdAndLessonIdAndDeletedAtIsNull(member.getMemberId(), lessonId).orElse(null);
+        if(homeworkRowList != null && putLessonReq.getHomeworks() == null ) {
+            needUpdate = Boolean.TRUE;
+        }
+        if(homeworkRowList == null && putLessonReq.getHomeworks() != null ) {
+            needUpdate = Boolean.TRUE;
+        }
+        if(homeworkRowList != null && putLessonReq.getHomeworks() != null) {
+            if(homeworkRowList.size() == putLessonReq.getHomeworks().size()) {
+                for(Homework homework : homeworkRowList) {
+                    for(HomeworkReq homeworkReq : putLessonReq.getHomeworks()) {
+                        if(!homework.isSameHomework(homeworkReq)) {
+                            needUpdate = Boolean.TRUE;
+                            break;
+                        }
+                    }
+                    if(needUpdate) {
+                        break;
+                    }
+                }
+            }
+            if(homeworkRowList.size() != putLessonReq.getHomeworks().size()) {
+                needUpdate = Boolean.TRUE;
+            }
+        }
+
+        // lesson은 무조건 업데이트 + 변경사항이 발생하면 숙제 업데이트 진행
+        lesson.update(putLessonReq);
+
+        if(needUpdate) {
+            homeworkRepository.deleteHomeworkByLessonId(lessonId);
+            List<Homework> homeworkInsertList = new ArrayList<>();
+            for (Participant participant : putLessonReq.getParticipants()) {
+                List<Homework> tempHomeworkList = putLessonReq.getHomeworks().stream()
+                        .map(homeworkReq -> HomeworkReq.staticToEntity(homeworkReq, participant.getMemberId(), lesson.getLessonId()))
+                        .collect(Collectors.toList());
+                homeworkInsertList.addAll(tempHomeworkList);
+            }
+            homeworkRepository.saveAll(homeworkInsertList);
+        }
 
         // Response
         return null;
@@ -148,7 +199,6 @@ public class LessonService {
 
         }
         lesson.updateMeta(patchLessonMetaRes);
-        System.out.println("size: " + homeworkList.size());
         homeworkRepository.saveAll(homeworkList);
